+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - Removed UUID related accessors on Entity and World.
 - `@PackedWeaver` marked as deprecated until performance issues have been resolved.
 - `EntityManager#isNew` removed.
+- Passive flag on systems removed.
 - **Fix**: No longer possible to accidentally add multiple systems of the same type.